pyproject.toml
[project]
name = "jobhunter"
version = "0.1.0"
description = "Job search automation platform"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy>=2.0,<3.0",
    "alembic>=1.13,<2.0",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
    "python-dotenv>=1.0,<2.0",
    "httpx>=0.27,<1.0",
    "pdfplumber>=0.11,<1.0",
    "click>=8.0,<9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.24,<1.0",
    "ruff>=0.8,<1.0",
    "mypy>=1.13,<2.0",
    "types-PyYAML>=6.0",
]
scraping = [
    "playwright>=1.49,<2.0",
]
ai = [
    "anthropic>=0.42,<1.0",
]
dashboard = [
    "streamlit>=1.40,<2.0",
]
all = [
    "playwright>=1.49,<2.0",
    "anthropic>=0.42,<1.0",
    "streamlit>=1.40,<2.0",
]

[project.scripts]
jobhunter = "jobhunter.cli:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["jobhunter*"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
